After insane investigations, I found some of the issues, some weird some not. r26-r30 have to be saved in stack and retrieved, else bad things can happen. Not sure about the rules of other regs. Something in input when -O2 or less failed. Seems for now if I changed the typedef of bool from uint8 to uint32 fixes the issue, but not sure if it hides the real problem or not.

+++ src/types.h
@@ -15,7 +15,7 @@ typedef ushort uint16;
 typedef unsigned long ulong;
 typedef ulong uint32;
 
-typedef	uint8 Boolean;
+typedef	uint32 Boolean;
 typedef Boolean	bool;
 
 #define TRUE	((Boolean)1)
